+++ src/mmop-worker.lisp
@@ -5,7 +5,8 @@
            make-worker-ready-v0
            start-node-v0
            start-node-v0-type
-           start-node-v0-recipe))
+           start-node-v0-recipe
+           make-start-node-success-v0))
 (in-package :monomyth/mmop-worker)
 
 (defstruct (worker-ready-v0 (:constructor make-worker-ready-v0 ()))
@@ -17,6 +18,13 @@
   "MMOP/0 start-node (note, the recipe should be deserialized)"
   (type (error "type must be set") :read-only t)
   (recipe (error "recipe must be set") :read-only t))
+
+(defstruct (start-node-success-v0
+            (:constructor make-start-node-success-v0 (type)))
+  "MMOP/0 start-node-success"
+  (type (error "type must be set") :read-only t))
+(defmethod create-frames ((message start-node-success-v0))
+  `(,*mmop-v0* "START-NODE-SUCCESS" ,(start-node-success-v0-type message)))
 
 (defun pull-worker-message (socket)
   "pulls down a message designed for the worker dealer socket and attempts to