+++ tests/mmop.lisp
@@ -106,4 +106,20 @@
             (is (rmq-node-recipe/source-queue got-res) (rmq-node-recipe/source-queue recipe))
             (is (rmq-node-recipe/dest-queue got-res) (rmq-node-recipe/dest-queue recipe))))))))
 
+(subtest "MMOP/0 worker-ready-success"
+  (let ((client-name (format nil "client-~a" (uuid:make-v4-uuid)))
+        (server-name (format nil "server-~a" (uuid:make-v4-uuid))))
+    (pzmq:with-context nil
+      (pzmq:with-sockets ((server :router) (client :dealer))
+        (pzmq:setsockopt server :identity server-name)
+        (pzmq:setsockopt client :identity client-name)
+        (pzmq:connect client "ipc://test.ipc")
+        (pzmq:bind server "ipc://test.ipc")
+
+        (send-msg client *mmop-v0* (mmop-w:make-start-node-success-v0 "TEST"))
+        (let ((res (mmop-m:pull-master-message server)))
+          (is-type res 'mmop-m:start-node-success-v0)
+          (is (mmop-m:start-node-success-v0-client-id res) client-name)
+          (is (mmop-m:start-node-success-v0-type res) "TEST"))))))
+
 (finalize)